Fixed 2 github minor issues and some warnings. https://github.com/amiq-consulting/fc4sc/issues/3 https://github.com/amiq-consulting/fc4sc/issues/5

+++ includes/fc4sc_intervals.hpp
@@ -44,7 +44,7 @@ namespace fc4sc {
     result_bins.push_back(new_bins[0]);
 
 
-    for (uint i = 1; i < new_bins.size(); ++i) {
+    for (size_t i = 1; i < new_bins.size(); ++i) {
 
       if(result_bins.back().second < new_bins[i].first)
         result_bins.push_back(new_bins[i]);
